csv was not working in runserver_multiple.py. Name csv conflicts python module csv. Renamed examples/csv to examples/csv_server.
